new folder/requirements (1).txt
cloudinary==1.41.0
sqlalchemy==2.0.23
social-post-api==1.5.0
python-dotenv==1.0.0
pillow==10.1.0